DataReader Almost Complete
The Room, Item, and Level Classes have been fleashed out along with the DataReader Systems. All that needs to be added is the final parts of the Item Data Reader and it is good to go. I wanted to say a version here as everythign is working and I would like to have the backup.